+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note that we start to track changes starting from v1.3.7.
 
 ### BoltDB
 - [Add support for loong64 arch](https://github.com/etcd-io/bbolt/pull/303).
+- Fix [the "Last" method might return no data due to not skipping the empty pages](https://github.com/etcd-io/bbolt/pull/341)
 
 ### CMD
 - [Open db file readonly in compact CLI command](https://github.com/etcd-io/bbolt/pull/292).